Add "who we're currently connected to" to the connected metric
This took quite a bit of wrangling, because peeraddr only works if we're connected,
and once the socket is closed (by us or anyone else), we're disconnected, so now
Target has to carry a lot more state.  It's actually a better design, IMO, it's
just I had to move a lot stuff around to get there.